+++ src/main/java/com/example/demo/repository/ISurveyQuestionRepo.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.SurveyQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ISurveyQuestionRepo extends JpaRepository<SurveyQuestion, Integer> {
+public interface ISurveyQuestionRepo extends JpaRepository<SurveyQuestion, Long> {
 }
